Reset with ongoing transaction (#128)
* Start with 'reset with ongoing transaction'.


* Start with database related things to store a transaction message from the queue in the database, so it can be sent after reboot if it is not sended yet.


* Implement database insert / remove / get functions.


* Add database base class in common library so the transactions in the queue can be stored in both versions (16 and 201). Put transaction in database as soon as it is received. Remove transaction message from database when reply is received. Some small improvements in for loop with copying / reference.


* Add evse id to reset callback. Store scheduled 'OnIdle' reset so it an be performed as soon as the charging has stopped.


* Add evseid to is_reset_allowed_callback


* Add documentation. Reset when all evse id's stopped charging. 


* 'initialize' sqlite db member in the constructor. Add documentation


* Make it possible to call the reset multiple times. Also call the availability handler on reset.


* Add two templates for conversion from message type to string.


* Change start() function and add optional boot reason.


* Add documentation about start function


* sqlite bind text fixes.


* Set flag to persist availability after reboot / reset


* Add persist to  as well


* Review comments about race conditions


* Fill transaction database after bootnotification is receveived and 'accepted'


* change database handler to shared ptr instead of unique ptr


* clang-format


---------

+++ lib/ocpp/v201/evse.cpp
@@ -33,8 +33,8 @@ EVSE Evse::get_evse_info() {
     return evse;
 }
 
-int32_t Evse::get_number_of_connectors() {
-    return this->id_connector_map.size();
+uint32_t Evse::get_number_of_connectors() {
+    return static_cast<uint32_t>(this->id_connector_map.size());
 }
 
 Everest::SteadyTimer& Evse::get_sampled_meter_values_timer() {
